dev_requirements.txt
asgiref
Django
django-admin-soft-dashboard==1.0.7
gunicorn
python-dotenv
whitenoise
pillow

django-author
django-birthday
django-bookmark-base
django-cors-headers
django-crispy-forms
django-dynamic-raw-id
django-filter
django-import-export
django-js-asset
django-mptt
django-phone-field
django-tables2
django-taggit
django-templated-mail
django-tinymce

